Add files via upload
The Alumni Network System is a web-based platform designed to connect former students of an educational institution, fostering a sense of community and facilitating networking opportunities. The system provides a central hub for alumni to stay engaged with their alma mater, share career updates, collaborate on professional opportunities, and participate in events.